split CustomCraftResultView
