+++ nekoyume/Assets/Resources/UI/Prefabs/UI_CustomCraftResultPopup.prefab
@@ -19448,6 +19448,7 @@ GameObject:
   - component: {fileID: 3320609653135551495}
   - component: {fileID: 2857340769885502676}
   - component: {fileID: 8439939730892565601}
+  - component: {fileID: 6104943854364415301}
   m_Layer: 5
   m_Name: UI_CustomCraftResultPopup
   m_TagString: Untagged
@@ -19520,6 +19521,19 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   tutorialTargets: []
   tutorialActions: 
+  view: {fileID: 6104943854364415301}
+--- !u!114 &6104943854364415301
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3244324374208210769}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 961bf0bae37747e0a6aa1f9dc592a861, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   itemView: {fileID: 1139161713777828630}
   itemNameText: {fileID: 8092519113964929652}
   baseStatText: {fileID: 183345820611038457}
