+++ nekoyume/Assets/_Scripts/UI/Module/CustomCraftResultView.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 961bf0bae37747e0a6aa1f9dc592a861
+timeCreated: 1726029415
